+++ contracts/entity-resolver.sol
@@ -7,14 +7,10 @@ import "./profiles/addr-resolver.sol";
 import "./profiles/text-resolver.sol";
 import "./profiles/text-list-resolver.sol";
 
-
 contract EntityResolver is AddrResolver, TextResolver, TextListResolver {
     function isAuthorised(bytes32 node) internal override view returns (bool) {
-        return getEntityId(msg.sender) == node;
-    }
-
-    function getEntityId(address entityAddress) public pure returns (bytes32) {
-        return keccak256(abi.encodePacked(entityAddress));
+        // The hash of the entity address must be the node
+        return keccak256(abi.encodePacked(msg.sender)) == node;
     }
 
     function supportsInterface(bytes4 interfaceID)