+++ lib/index.ts
@@ -2,7 +2,7 @@ import * as EntityResolver from "./entity-resolver.json"
 import * as Process from "./process.json"
 import * as Namespace from "./namespace.json"
 
-import { ContractTransaction } from "ethers"
+import { ContractTransaction, utils } from "ethers"
 import { BigNumber } from "ethers/utils"
 
 ///////////////////////////////////////////////////////////////////////////////
@@ -22,9 +22,6 @@ export type EntityResolverContractMethods = {
     /** Whether the resolver supports an interface */
     supportsInterface(interfaceID: string): Promise<boolean>
 
-    /** Get the entity ID for a given Ethereum address */
-    getEntityId(entityAddress: string): Promise<string>
-
     /** Get the address associated with the given node */
     addr(node: string): Promise<string>
     /** Sets the address for the given node */
@@ -32,61 +29,69 @@ export type EntityResolverContractMethods = {
 
     /**
      * Returns the text associated with an ENS node and key.
-     * @param entityId The ENS node to query.
+     * @param hashedEntityAddress The ENS node to query.
      * @param key The key to retrieve.
      * @return The record's text.
      */
-    text(entityId: string, key: string): Promise<string>
+    text(hashedEntityAddress: string, key: string): Promise<string>
     /**
      * Sets the text of the ENS node and key.
      * May only be called by the owner of that node in the ENS registry.
-     * @param entityId The ENS node to modify.
+     * @param hashedEntityAddress The ENS node to modify. @see ensHashAddress()
      * @param key The key to modify.
      * @param value The text to store.
      */
-    setText(entityId: string, key: string, value: string): Promise<ContractTransaction>
+    setText(hashedEntityAddress: string, key: string, value: string): Promise<ContractTransaction>
 
     /**
      * Returns the list associated with an ENS node and key.
-     * @param entityId The ENS node to query.
+     * @param hashedEntityAddress The ENS node to query.
      * @param key The key of the list.
      * @return The list array of values.
      */
-    list(entityId: string, key: string): Promise<string[]>
+    list(hashedEntityAddress: string, key: string): Promise<string[]>
     /**
      * Returns the text associated with an ENS node, key and index.
-     * @param entityId The ENS node to query.
+     * @param hashedEntityAddress The ENS node to query.
      * @param key The key of the list.
      * @param index The index within the list to retrieve.
      * @return The list entry's text value.
      */
-    listText(entityId: string, key: string, index: number): Promise<string>
+    listText(hashedEntityAddress: string, key: string, index: number): Promise<string>
     /**
      * Sets the text of the ENS node, key and index.
      * May only be called by the owner of that node in the ENS registry.
-     * @param entityId The ENS node to modify.
+     * @param hashedEntityAddress The ENS node to modify. @see ensHashAddress()
      * @param key The key of the list to modify.
      * @param index The index of the list to set.
      * @param value The text to store.
      */
-    setListText(entityId: string, key: string, index: number, value: string): Promise<ContractTransaction>
+    setListText(hashedEntityAddress: string, key: string, index: number, value: string): Promise<ContractTransaction>
     /**
      * Appends a new value on the given ENS node and key.
      * May only be called by the owner of that node in the ENS registry.
-     * @param entityId The ENS node to modify.
+     * @param hashedEntityAddress The ENS node to modify. @see ensHashAddress()
      * @param key The key of the list to modify.
      * @param value The text to store.
      */
-    pushListText(entityId: string, key: string, value: string): Promise<ContractTransaction>
+    pushListText(hashedEntityAddress: string, key: string, value: string): Promise<ContractTransaction>
     /**
      * Removes the value on the ENS node, key and index.
      * May only be called by the owner of that node in the ENS registry.
      * Note: This may cause items to be arranged in a different order.
-     * @param entityId The ENS node to modify.
+     * @param hashedEntityAddress The ENS node to modify. @see ensHashAddress()
      * @param key The key of the list to modify.
      * @param index The index to remove.
      */
-    removeListIndex(entityId: string, key: string, index: number): Promise<ContractTransaction>
+    removeListIndex(hashedEntityAddress: string, key: string, index: number): Promise<ContractTransaction>
+}
+
+/**
+ * Hashes the address of an Entity to interact with the Entity Resolver contract
+ * @param address
+ */
+export function ensHashAddress(address: string): string {
+    return utils.keccak256(address)
 }
 
 ///////////////////////////////////////////////////////////////////////////////